depro_full/backend/requirements.txt
fastapi
uvicorn
python-multipart
gitpython
openai
python-dotenv
google-genai
ollama
boto3
botocore
requests
paramiko
cryptography
langgraph
langchain-core
typing-extensions
pydantic
py-algorand-sdk
algokit-utils
puyapy
algorand-sdk
algorand-python